requirements_simple.txt
# Simplified FastAPI requirements for Windows compatibility
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
websockets>=11.0
aiohttp>=3.8.0
httpx>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
coloredlogs>=15.0.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.0
